refactor: extract referral reward to its own component
This will allow the FormAuctionReferral component to be used without an
auction
